second enhancement finished
