+++ src/test.js
@@ -1,0 +1,4 @@
+const util = require('./util');
+
+
+util.bump('1.0.0', 'major');
